requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
filetype==1.2.0
pypdf==5.5.0
sqlparse==0.5.3
